requirements.txt
asgiref==3.4.1
Django==3.2.8
django-crispy-forms==1.13.0
django-phonenumber-field==5.2.0
phonenumbers==8.12.34
pytz==2021.3
sqlparse==0.4.2
